backend/tfall/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.1.2
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.4
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
whitenoise==6.2.0
